3.5更新内容
【功能新增】
1.新增后端文件存储（对象存储）模块 platform-boot-starter-oss
2.新增系统支持模块platform-support
3.前端集成文件上传组件vue-simple-uploader
4.前端改造富文本编辑器wangEditor
4.实现通知公告功能
【优化调整】
1.文件下载实现方案优化，原方案采用blob方式，小文件无问题，大文件需整个文件下载完成后浏览器才会显示，下载期间看上去无响应，影响体验，调整为url直连方式，点击下载浏览器立即显示开始下载并显示进度。
2.调整代码生成模板，新增、修改、查看视图默认宽度由百分比固化为像素
【bug修复】
1.修复参照视图模板if标签未闭合的bug
2.移除vo代码模板多余的括号
3.实体列表固化排序逻辑，先按模块升序，后按排序号升序，避免默认按排序号导致多模块实体混在一起的杂乱情况

+++ platform-system/src/main/java/tech/abc/platform/system/service/impl/UserServiceImpl.java
@@ -347,7 +347,7 @@ public class UserServiceImpl extends BaseServiceImpl<UserMapper, User> implement
 
     @Override
     public void checkPermission(String id, String permissionCode) {
-        int count = userMapper.checkPermission(id, permissionCode);
+        long count = userMapper.checkPermission(id, permissionCode);
         if (count == 0) {
             throw new CustomException(PermissionItemExceptionEnum.NO_PERMISSION);
         }
